Fix left-shift signed overflows (UB) (#62)
* Fix left-shift signed overflows (UB)

In both cases pm may be negative, and left shifting a negative value is
signed overflow (undefined behavior). The destination is 16-bit, and
8-bit pm is always left-shifted by 8, shifting any sign extension of pm
beyond the destination. Therefore its signedness plays no practical role
in these operations.

These shift operands are first promoted to int, and so it's still signed
left-shifts of non-negative operands. 32-and 64-bit host have plenty of
overhead. On 16-bit hosts this shifts into the sign bit, and no farther,
a well-defined special case in C++ (CWG 1457).

Removing the explicit int8_t casts isn't necessary, but they no longer
make sense with this change.

I discovered these via Undefined Behavior Sanitizer on x16emu.

* Fix signed overflow in roundtrip_fp

If value is negative, the final left shift is undefined. To avoid this
replace both shifts with an equivalent mask.

* Fix signed overflow in compute_phase_step

If lfo_raw_pm is negative, the left shift is undefined. Convert to
unsigned for the shift so that it's defined. It's still unsigned when
accumulated into delta, but this doesn't change the result.